Fix internet connection for Exchanges

+++ TMS.Net07.Lesson1/SpaceWeb/EfStuff/SeedExtension.cs
@@ -193,6 +193,11 @@ namespace SpaceWeb.EfStuff
         private static void SetDefaultExchangeRateToUsdCurrent(IServiceProvider services)
         {
             var currencyService = services.GetService<ICurrencyService>();
+
+            if (!currencyService.CheckInternetConnection())
+            {
+                return;
+            }
             var exchRateToUsdCurrentRepository = services.GetService<ExchangeRateToUsdCurrentRepository>();
 
             currencyService.DeleteCurrentExchRatesFromDb(exchRateToUsdCurrentRepository);
